Move Material.h to Materials folder
