+++ GameProject/Source/Game.cpp
@@ -2,7 +2,7 @@
 
 #include "Goknar/Scene.h"
 #include "Goknar/Model/StaticMesh.h"
-#include "Goknar/Material.h"
+#include "Goknar/Materials/Material.h"
 #include "Goknar/Managers/ResourceManager.h"
 #include "Goknar/Camera.h"
 #include "Goknar/Managers/CameraManager.h"
