+++ Goknar/Source/Goknar/Helpers/SceneParser.cpp
@@ -9,18 +9,18 @@
 #include "Goknar/Managers/CameraManager.h"
 #include "Goknar/Engine.h"
 #include "Goknar/Scene.h"
-#include "Goknar/Material.h"
-#include "Goknar/Model/MeshUnit.h"
-#include "Goknar/Model/StaticMesh.h"
 #include "Goknar/Components/MeshComponent.h"
 #include "Goknar/IO/ModelLoader.h"
 #include "Goknar/ObjectBase.h"
+#include "Goknar/Materials/Material.h"
 #include "Goknar/Renderer/Shader.h"
 #include "Goknar/Renderer/Texture.h"
+
 
 #include "Goknar/Managers/ResourceManager.h"
 
 #include "Goknar/Model/DynamicMesh.h"
+#include "Goknar/Model/MeshUnit.h"
 #include "Goknar/Model/StaticMesh.h"
 #include "Goknar/Model/SkeletalMesh.h"
