+++ Goknar/Source/Goknar/IO/ModelLoader.cpp
@@ -6,7 +6,7 @@
 #include "Goknar/Contents/Image.h"
 #include "Goknar/Engine.h"
 #include "Goknar/GoknarAssert.h"
-#include "Goknar/Material.h"
+#include "Goknar/Materials/Material.h"
 #include "Goknar/Managers/ResourceManager.h"
 #include "Goknar/Model/SkeletalMesh.h"
 #include "Goknar/Model/StaticMesh.h"
