+++ Goknar/Source/Goknar/Lights/ShadowManager/ShadowManager.cpp
@@ -4,10 +4,10 @@
 
 #include "Goknar/Application.h"
 #include "Goknar/Camera.h"
+#include "Goknar/Materials/Material.h"
 #include "Goknar/Managers/CameraManager.h"
 #include "Goknar/Engine.h"
 #include "Goknar/Log.h"
-#include "Goknar/Material.h"
 #include "Goknar/Renderer/Renderer.h"
 #include "Goknar/Scene.h"
 #include "Goknar/Renderer/Shader.h"
