+++ Goknar/Source/Goknar/Materials/MaterialInstance.cpp
@@ -1,0 +1,36 @@
+ #include "pch.h"
+
+#include "MaterialInstance.h"
+
+MaterialInstance::MaterialInstance()
+{
+}
+
+MaterialInstance::MaterialInstance(Material* parent) :
+	Material(parent)
+{
+
+}
+
+MaterialInstance::~MaterialInstance()
+{
+}
+
+void MaterialInstance::Create(const Material* parent)
+{
+}
+
+void MaterialInstance::PreInit()
+{
+
+}
+
+void MaterialInstance::Init()
+{
+
+}
+
+void MaterialInstance::PostInit()
+{
+
+}
