+++ Goknar/Source/Goknar/Model/Mesh.cpp
@@ -2,7 +2,7 @@
 
 #include "Mesh.h"
 
-#include "Material.h"
+#include "Materials/Material.h"
 #include "MeshUnit.h"
 
 Mesh::Mesh()
