+++ Goknar/Source/Goknar/Model/SkeletalMeshInstance.cpp
@@ -6,8 +6,8 @@
 #include <execution>
 
 #include "Goknar/Engine.h"
-#include "Goknar/Material.h"
 #include "Goknar/Components/SocketComponent.h"
+#include "Goknar/Materials/Material.h"
 #include "Goknar/Model/SkeletalMesh.h"
 #include "Goknar/Renderer/Shader.h"
 #include "Goknar/Renderer/Renderer.h"
