+++ Goknar/Source/Goknar/Renderer/ShaderBuilder.cpp
@@ -8,9 +8,9 @@
 #include "Goknar/Scene.h"
 #include "Goknar/Lights/DirectionalLight.h"
 #include "Goknar/IO/IOManager.h"
+#include "Goknar/Materials/Material.h"
 #include "Goknar/Model/MeshUnit.h"
 #include "Goknar/Model/SkeletalMesh.h"
-#include "Goknar/Material.h"
 #include "Goknar/Lights/PointLight.h"
 #include "Goknar/Lights/SpotLight.h"
 #include "Goknar/Renderer/Texture.h"
